social_network/requirements.txt
django==5.2
djangorestframework==3.14.0
pillow==11.1.0
psycopg2-binary==2.9.10
